requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.5
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
pydantic==1.10.11
pydantic_core==0.42.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.23.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
python-multipart
gunicorn